Extract ImportableDiscovery: HTTP browser + adoption flow
Fifth (final) PR in the device-state-monitor split arc. No direct
legacy sibling — this is the HTTP browser + DashboardImportDiscovery
layer the new dashboard added on top of the legacy mDNS browser.

**``importable.py``** — new ``ImportableDiscovery`` class:

* ``setup`` — constructs the upstream ``DashboardImportDiscovery``
  (callback wires through ``self._on_import_update``). Called
  from :meth:`MdnsSource.start`.
* ``browser_callback`` — forwards esphomelib browser events to
  the upstream ``DashboardImportDiscovery.browser_callback`` so
  the dispatch closure in mdns.py no longer needs to know about
  it.
* ``probe_device`` / ``revisit_importable`` /
  ``revisit_all_importables`` / ``get_importable_devices`` —
  the public discovery / adoption surface, reaching back
  through ``self._monitor._mdns.zeroconf`` / ``._apply_service_info``
  / ``._resolve_and_apply`` for the cache-hit fast path.
* ``on_http_service_state_change`` — was
  ``_on_http_service_state_change``; the HTTP-browser callback
  the dispatch closure invokes for ``_http._tcp.local.`` events.
* ``_resolve_and_apply_http`` / ``_apply_http_service_info`` /
  ``_seed_http_url_from_cache`` — HTTP-service-info plumbing.
* ``_has_importable`` / ``_refire_importable_for`` —
  importable-state inspection / re-emit helpers.
* ``_on_import_update`` / ``_build_adoptable`` — bridge from
  upstream ``DiscoveredImport`` to our ``AdoptableDevice`` model.

The ``_import_discovery`` handle moves to ImportableDiscovery.

**``mdns.py``**: the dispatch closure routes esphomelib events to
``self._on_esphomelib_service_state_change`` AND
``importable.browser_callback`` (replacing the previous local
``DashboardImportDiscovery`` construction), and HTTP events to
``importable.on_http_service_state_change``. ``DashboardImportDiscovery``
import dropped.

**``controller.py``** wiring:

* ``__init__`` builds ``self._importable = ImportableDiscovery(self)``
  alongside ``self._mdns`` / ``self._ping``.
* The ``_import_discovery`` field on the monitor goes away.
* Public methods (``probe_device``, ``revisit_importable``,
  ``revisit_all_importables``, ``get_importable_devices``)
  become one-line delegates to ``self._importable.*``.
* All HTTP / adoption private methods removed: ``_on_http_service_state_change``,
  ``_resolve_and_apply_http``, ``_apply_http_service_info``,
  ``_has_importable``, ``_refire_importable_for``,
  ``_seed_http_url_from_cache``, ``_on_import_update``,
  ``_build_adoptable``.

**Test redirects** (mechanical):

* Patches against ``state_monitor_module.AsyncServiceInfo`` for
  HTTP tests redirect to ``importable_module.AsyncServiceInfo``;
  the esphomelib-dispatch tests (already on ``mdns_module``)
  unchanged.
* ``mdns_module.DashboardImportDiscovery`` patches redirect to
  ``importable_module.DashboardImportDiscovery``.
* ``monitor._import_discovery`` / ``monitor._on_import_update`` /
  ``monitor._apply_http_service_info`` / ``monitor._has_importable``
  / ``monitor._refire_importable_for`` /
  ``monitor._seed_http_url_from_cache`` /
  ``monitor._on_http_service_state_change`` /
  ``monitor._resolve_and_apply_http`` / ``monitor._build_adoptable``
  → ``monitor._importable.*``.
* ``state_monitor_module.device_name_from_service`` →
  ``helpers_module.device_name_from_service`` (the controller
  no longer needs to import the helper).
* Probe-device patches against
  ``controller.AsyncServiceInfo`` →
  ``importable.AsyncServiceInfo``.
* ``__new__``-built test fixtures now wire ``monitor._importable
  = ImportableDiscovery(monitor)`` before ``monitor._mdns``.

controller.py: 755 → 528 lines.

End-of-arc shape:

| File | Lines |
|---|---:|
| ``__init__.py`` | 25 |
| ``_state.py`` | 41 |
| ``controller.py`` | 528 |
| ``helpers.py`` | 208 |
| ``importable.py`` | 286 |
| ``mdns.py`` | 453 |
| ``ping.py`` | 245 |
| ``shared.py`` | 156 |

Original 1638-line ``_device_state_monitor.py`` is now 8
modules totalling 1942 lines (overhead from per-module
docstrings + imports). Every file under the 800-line soft
cap; per-concern modules testable in isolation.
